@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 30
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 31
     defaultConfig {
         applicationId "kr.azazel.barcode"
-        minSdkVersion 16
-        targetSdkVersion 30
-        versionCode 32
-        versionName "3.2"
+        minSdkVersion 24
+        targetSdkVersion 31
+        versionCode 38
+        versionName "3.8"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
     }
@@ -36,21 +35,28 @@ dependencies {
     implementation project(':framework')
     implementation project(':tedbottompicker')
 
-    implementation 'com.android.support:design:24.2.1'
-    implementation 'com.android.support:recyclerview-v7:24.2.1'
-    implementation 'com.android.support:cardview-v7:24.2.1'
-    implementation 'com.google.android.gms:play-services-vision:9.6.1'
+    implementation 'androidx.core:core:1.8.0'
+    implementation 'com.google.android.material:material:1.5.0'
 
     implementation 'com.jakewharton:butterknife:7.0.1'
     implementation 'com.squareup.picasso:picasso:2.5.2'
 
-    implementation 'com.journeyapps:zxing-android-embedded:3.4.0'
+    implementation 'com.google.mlkit:barcode-scanning:17.0.2'
+//    implementation 'com.google.android.gms:play-services-mlkit-barcode-scanning:18.0.0'
+
+    implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
 
     implementation 'com.theartofdev.edmodo:android-image-cropper:2.3.+'
     implementation 'com.github.rey5137:material:1.2.4'
 
-    implementation 'com.google.firebase:firebase-ads:9.6.1'
-    implementation 'com.google.firebase:firebase-core:9.6.1'
+    // Import the BoM for the Firebase platform
+    implementation platform('com.google.firebase:firebase-bom:30.1.0')
+
+    // Declare the dependencies for the Crashlytics and Analytics libraries
+    // When using the BoM, you don't specify versions in Firebase library dependencies
+    implementation 'com.google.android.gms:play-services-ads:21.0.0'
+    implementation 'com.google.firebase:firebase-crashlytics'
+    implementation 'com.google.firebase:firebase-analytics'
 
     implementation 'commons-io:commons-io:2.4'
     implementation "com.facebook.fresco:fresco:1.2.0"
@@ -61,3 +67,5 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.13.3'
 }
 apply plugin: 'com.google.gms.google-services'
+// Apply the Crashlytics Gradle plugin
+apply plugin: 'com.google.firebase.crashlytics'
