@@ -3,8 +3,8 @@ package kr.azazel.barcode;
 public class AzAppConfig {
     public static final String PUBLISHER = "Azazel";
     public static final String NAME = "AzBarcode";
-    public static final String VERSION = "3.2";
-    public static final int VERSION_CODE = 32;
+    public static final String VERSION = "3.8";
+    public static final int VERSION_CODE = 38;
 
     public static final boolean LOG_ENABLED = false;
     public static final boolean FILE_LOG_ENABLED = false;
